Add process_pa route and refactor/cleanup.
